feat: updates ui

+++ agenda/presentation/src/main/java/com/juandgaines/agenda/presentation/agenda_item/AgendaItemScreen.kt
@@ -8,9 +8,10 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.foundation.layout.size
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
@@ -155,8 +156,7 @@ fun AgendaItemScreen(
                             verticalAlignment = androidx.compose.ui.Alignment.CenterVertically,
                             modifier = Modifier
                                 .fillMaxWidth()
-                                .wrapContentHeight(),
-                            horizontalArrangement = Arrangement.SpaceBetween
+                                .size(56.dp),
                         ){
                             Text(
                                 text = state.title,
@@ -227,7 +227,7 @@ fun AgendaItemScreen(
         ){
             Column (
                 modifier = Modifier
-                    .fillMaxWidth()
+                    .fillMaxSize()
                     .padding(16.dp),
                 verticalArrangement = Arrangement.spacedBy(24.dp)
             ){
@@ -283,15 +283,18 @@ fun AgendaItemScreen(
 
                 Spacer(modifier = Modifier.weight(1f))
 
-                Text(
-                    text = stringResource(id = R.string.delete_item, agendaItemName).uppercase(),
-                    style = MaterialTheme.typography.titleSmall,
-                    color = TaskyGray,
-                    textAlign = TextAlign.Center,
-                    modifier = Modifier
-                        .padding(16.dp)
-                        .fillMaxWidth()
-                )
+                if(state.isNew.not()){
+                    Text(
+                        text = stringResource(id = R.string.delete_item, agendaItemName).uppercase(),
+                        style = MaterialTheme.typography.titleSmall,
+                        color = TaskyGray,
+                        textAlign = TextAlign.Center,
+                        modifier = Modifier
+                            .padding(16.dp)
+                            .fillMaxWidth()
+                            .clickable { onAction(AgendaItemAction.Delete) }
+                    )
+                }
             }
 
         }
